server/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
requests>=2.31,<3.0
psycopg[binary]>=3.2,<3.4
python-dotenv>=1.0,<2.0
reportlab>=4.2,<4.3
python-docx>=1.1,<1.2
